+++ tests/helpers/wait-for-max-height.js
@@ -1,0 +1,5 @@
+import { find, waitUntil } from '@ember/test-helpers';
+
+export default function waitForMaxHeight(selector, maxHeight) {
+  return waitUntil(() => find(selector).style.maxHeight === maxHeight);
+}